Funcion modificada para Vic

+++ src/java/model/mdo/DropboxPersistence.java
@@ -167,20 +167,16 @@ public final class DropboxPersistence implements FilePersistence {
         return file;
     }
 
-    public File descargarArchivoHTML(String ruta) throws IOException, DbxException {
-        String nombre = "preview.html";
+    public File descargarArchivoHTML(String ruta, String nombre) throws IOException, DbxException {
         String appRoot = ServletActionContext.getRequest().getServletContext().getRealPath("/");
-        String token = ruta.split("/")[1];
-        String tempFile = appRoot + ruta + "/" + nombre;
-        File temp = new File(tempFile);
-        FileUtils.writeStringToFile(temp, "");
-
-        try (FileOutputStream fos = new FileOutputStream(temp)) {
-            client.files().download(ruta + "/" + nombre).download(fos);
+        File temp = new File(appRoot + ruta + "/" + nombre);
+        if(!temp.exists())
+        {
+            FileUtils.writeStringToFile(temp, "");
+            try (FileOutputStream fos = new FileOutputStream(temp)) {
+                client.files().download(ruta + "/" + nombre).download(fos);
+            }
         }
-
-        //String file = FileUtils.readFileToString(temp);
-        //FileUtils.deleteDirectory(new File(appRoot + token));
         return temp;
     }
